Conflicts:
	CatProducer/python/catTools_cff.py

+++ CatAnalyzer/prod/TTLJ/pass1.2_cleanup.py
@@ -6,36 +6,11 @@ from os.path import join as pathjoin
 from multiprocessing import Pool, cpu_count
 import imp
 
-prefix = "ntuple"
-outDir = "pass1"
+from CATTools.CommonTools.condorTools import jobStatus
+from CATTools.CommonTools.haddsplit import haddsplit
 
-def hadd2(destDir, inputFiles):
-    ## First split files by prefix
-    inputFiles.sort(key = lambda x : int(x.split('_')[-1][:-5]))
-
-    ## estimate total size and merge not to exceed 1Gbytes
-    totalSize = sum([os.stat(f).st_size for f in inputFiles])
-    fSize = 1024.*1024*1024 # 1GB
-    nOutput = int(totalSize/fSize)+1
-
-    if nOutput == 1:
-        cmd = 'hadd -f %s.root ' % destDir
-        cmd += ' '.join(inputFiles)
-        os.system(cmd)
-    else:
-        nTotalFiles = len(inputFiles)
-        nFiles = nTotalFiles/nOutput
-
-        for i in range(nOutput):
-            print i
-            cmd = 'hadd -f %s_%d.root ' % (destDir, i)
-            if i == nOutput-1: cmd += ' '.join(inputFiles[i*nFiles:])
-            else: cmd += ' '.join(inputFiles[i*nFiles:(i+1)*nFiles])
-            os.system(cmd)
-    for x in inputFiles: os.system("rm -f %s" % (x))
-
-def cleanup(d):
-    d = d.rstrip('/')
+def cleanup(srcDir, outDir):
+    d = srcDir.rstrip('/')
     if not isdir(d): return 
 
     nqueue = 0
@@ -46,12 +21,10 @@ def cleanup(d):
 
     rootFiles = []
     for x in os.listdir(d):
-        if not x.endswith('.root') or not x.startswith(prefix): continue
+        if not x.endswith('.root'): continue
         x = pathjoin(d, x)
         if os.stat(x).st_size <= 200: continue
         rootFiles.append(x)
-
-    print len(rootFiles), nqueue
 
     if len(rootFiles) == nqueue:
         print
@@ -59,10 +32,11 @@ def cleanup(d):
         print "*", d, "is complete. Merge and clean up"
         print "*"*40
         print
-        hadd2(d, rootFiles)
-        os.system("rm -f %s/job.tar.gz" % d)
-        os.system("tar czf %s/log.tgz %s/*.log %s/*.err %s/*.txt" % (d, d, d, d))
-        os.system("rm -f %s/*.log %s/*.err %s/*.txt" % (d, d, d))
+        isOK = haddsplit(srcDir, rootFiles, outDir)
+        if isOK:
+            os.system("rm -f %s/job.tar.gz" % d)
+            os.system("tar czf %s/log.tgz %s/*.log %s/*.err %s/*.txt" % (d, d, d, d))
+            os.system("rm -f %s/*.log %s/*.err %s/*.txt" % (d, d, d))
     else:
         print
         print "+"*40
@@ -70,89 +44,61 @@ def cleanup(d):
         print "+"*40
         print
 
-def jobStatus(d):
-    if not os.path.exists("%s/condor.log" % d): return [], [], []
-    jobs = {}
-    lines = [x.strip() for x in open("%s/condor.log" % d).readlines()]
-    for i, l in enumerate(lines):
-        if 'Job submitted' in l:
-            section = int(l.split()[1][1:-1].split('.')[1])
-            jobs[section] = 'SUBMIT'
-        elif 'Job terminated' in l:
-            section = int(l.split()[1][1:-1].split('.')[1])
-            retVal = int(lines[i+1].split()[-1][:-1])
-            if retVal == 0: jobs[section] = 'DONE'
-            else: jobs[section] = 'ERROR %d' % retVal
-
-    jobsSub, jobsDone, jobsErr = [], [], []
-    for section in jobs:
-        if jobs[section] == 'SUBMIT': jobsSub.append(section)
-        elif jobs[section] == 'DONE': jobsDone.append(section)
-        else: jobsErr.append(section)
-    return jobsSub, jobsDone, jobsErr
-
 if __name__ == '__main__':
-    pool = Pool(cpu_count())
+    srcBase = "pass1"
+    outBase = "pass2"
 
     nSub = 0
     jobsFinished, jobsFailed = [], []
-    for sample in os.listdir(outDir):
-        sample = pathjoin(outDir, sample, 'nominal')
-        if not isdir(sample): continue
+    for sample, dirName, files in os.walk(srcBase):
+        if '.create-batch' not in files or 'job.tar.gz' not in files: continue
+        if 'condor.log' not in files:
+            print sample, " incomplete submission"
+            continue
 
         stat = jobStatus(sample)
-        if len(stat[0])+len(stat[1])+len(stat[2]) == 0: continue
+        if len(stat[0])+len(stat[1])+len(stat[2]) == 0:
+            print sample, " without jobs"
+            continue
 
         nSub += len(stat[0])
         if len(stat[0])+len(stat[1])+len(stat[2]) == 0: continue
         elif len(stat[0])+len(stat[2]) == 0: jobsFinished.append(sample)
         elif len(stat[0]) == 0 and len(stat[2]) > 0: jobsFailed.append((sample, stat[2][:]))
 
-    for sample in [pathjoin(outDir, x) for x in os.listdir(outDir) if isdir(pathjoin(outDir, x))]:
-        for subsample in [pathjoin(sample, x) for x in os.listdir(sample) if isdir(pathjoin(sample, x))]:
-            for syst in [pathjoin(subsample, x) for x in os.listdir(subsample) if isdir(pathjoin(subsample, x))]:
-                stat = jobStatus(syst)
-                if len(stat[0])+len(stat[1])+len(stat[2]) == 0: continue
-
-                nSub += len(stat[0])
-                if len(stat[0])+len(stat[1])+len(stat[2]) == 0: continue
-                elif len(stat[0])+len(stat[2]) == 0: jobsFinished.append(syst)
-                elif len(stat[0]) == 0 and len(stat[2]) > 0: jobsFailed.append((syst, stat[2][:]))
-
-    outFiles = []
+    pool = Pool(cpu_count())
     for sample in jobsFinished:
-        pool.apply_async(cleanup, [sample])
-        outFiles.append(sample+'.root')
-
+        outDir = outBase+'/'+sample.split('/',1)[-1]
+        pool.apply_async(cleanup, [sample, outDir])
     pool.close()
     pool.join()
 
     if len(jobsFailed) > 0:
         print "@@ There are some failed jobs"
         print "@@ Please resubmit following jobs"
-        if os.path.exists("%s/resubmit" % outDir): os.system("rm -rf %s/resubmit" % outDir)
-        os.makedirs("%s/resubmit" % outDir)
-        fsubmit = open("%s/resubmit/submit.sh" % outDir, "w")
+        if os.path.exists("%s/resubmit" % srcBase): os.system("rm -rf %s/resubmit" % srcBase)
+        os.makedirs("%s/resubmit" % srcBase)
+        fsubmit = open("%s/resubmit/submit.sh" % srcBase, "w")
         for sample, jobs in jobsFailed:
             print sample, jobs
             files = []
             for job in jobs:
-                os.system("rm -f %s/%s_%03d.root" % (sample, prefix, job))
+                os.system("rm -f %s/*_%03d.root" % (sample, job))
                 process = imp.load_source("process", "%s/job_%03d_cfg.py" % (sample, job)).process
                 for file in process.source.fileNames:
                     files.append(file)
-            prefix = sample.replace("%s/" % outDir, "").replace("/", "_")
-            frerun = open("%s/resubmit/%s.txt" % (outDir, prefix), "w")
+            prefix = sample.replace("%s/" % srcBase, "").replace("/", "_")
+            frerun = open("%s/resubmit/%s.txt" % (srcBase, prefix), "w")
             for f in files: print>>frerun, f
             frerun.close()
-            frerun = open("%s/resubmit/%s_cfg.py" % (outDir, prefix), "w")
+            frerun = open("%s/resubmit/%s_cfg.py" % (srcBase, prefix), "w")
             frerun.write(process.dumpPython())
             frerun.close()
 
             print>>fsubmit, ("create-batch --jobName {0} --fileList {0}.txt --cfg {0}_cfg.py --maxFiles 1".format(prefix))
 
         fsubmit.close()
-        os.system("chmod +x %s/resubmit/submit.sh" % outDir)
+        os.system("chmod +x %s/resubmit/submit.sh" % srcBase)
 
     if nSub > 0:
         print "@@ Done, wait for the %d jobs to finish" % nSub
